requirements.txt
django-crontab
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
flower==2.0.1  # Optional for monitoring